Cleanup src/highcharts content and public API
- Coloring stuff moved to base/interfaces
- Proptypes moved to src/proptypes
