+++ libs/sdk-ui/src/highcharts/chart/highcharts/plugins/test/dataLabelsColors.spec.ts
@@ -2,7 +2,8 @@
 import { isWhiteNotContrastEnough } from "../dataLabelsColors";
 
 import { getRgbString } from "../../../../utils/color";
-import { DEFAULT_COLOR_PALETTE, IColorPaletteItem } from "../../../../Config";
+import { DEFAULT_COLOR_PALETTE } from "../../../../Config";
+import { IColorPaletteItem } from "../../../../../base/interfaces/Colors";
 
 describe("dataLabelsColors", () => {
     describe("isWhiteNotContrastEnough", () => {
