+++ libs/sdk-ui/src/highcharts/chart/test/ChartTransformation.spec.tsx
@@ -7,8 +7,9 @@ import ChartTransformation from "../ChartTransformation";
 import * as fixtures from "../../../../__mocks__/fixtures";
 import { TOP } from "../legend/PositionTypes";
 import HighChartsRenderer from "../HighChartsRenderer";
-import { IChartConfig, IColorPaletteItem } from "../../Config";
+import { IChartConfig } from "../../Config";
 import { getRgbString } from "../../utils/color";
+import { IColorPaletteItem } from "../../../base/interfaces/Colors";
 
 describe("ChartTransformation", () => {
     const defaultProps = {
