+++ libs/sdk-ui/src/internal/interfaces/Colors.ts
@@ -1,7 +1,7 @@
 // (C) 2019 GoodData Corporation
 import * as MappingHeader from "../../base/interfaces/MappingHeader";
 import { IColor, IColorItem } from "@gooddata/sdk-model";
-import { IColorPaletteItem, IColorAssignment } from "../../highcharts";
+import { IColorPaletteItem, IColorAssignment } from "../../base/interfaces/Colors";
 
 export interface IColoredItem {
     colorItem: IColorItem;
