+++ libs/sdk-ui/src/proptypes/Filters.ts
@@ -1,4 +1,5 @@
 // (C) 2007-2018 GoodData Corporation
-import { FiltersPropTypesShape } from "../highcharts";
+
+import { FiltersPropTypesShape } from "./execution";
 
 export const FiltersPropType: any = FiltersPropTypesShape;
